Fix indentation
One of the biggest irregularities here was stuff like this:

    some_var
      |> piped_through_thing_1
      |> and_also_through_thing_2

The convention in Elixir is to indent the successive lines that start
with the pipe, `|>`, to the same indentation level as the first line,
like this:

    some_var
    |> piped_through_thing_1
    |> and_also_through_thing_2

I understand wanting to indent these the other way, but I think it's
important to get used to the convention, as it's what you're going to
see when you read Elixir code out on the web.

Other than that, I think I saw a couple things that indicate that you
might not fully understand how to read `case` statements in Elixir.
Refresh your memory on that!
https://elixirschool.com/en/lessons/basics/control-structures/